% Auto-generated: do not edit by hand
\name{''UpsetJS}

\alias{''UpsetJS}

\title{UpsetJS component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''UpsetJS(id=NULL, data=NULL, height=NULL, interaction=NULL,
selection=NULL, setLabelSize=NULL, setName=NULL, theme=NULL,
title=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{data}{List of lists containing elements 'name', 'sets'.
those elements have the following types:
  - name (character; required)
  - sets (list of characters; required)s. }

\item{height}{Numeric. }

\item{interaction}{Character. }

\item{selection}{Named list. }

\item{setLabelSize}{Character. }

\item{setName}{Character. }

\item{theme}{Character. }

\item{title}{Character. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

